refactor(builtins): fold the compiled path's arms into dispatch_pure
call_pure_builtin is gone; its twenty-nine arms now live in the shared
dispatcher and the compiled path calls that instead.

This is the safe half of the extraction. Instrumenting call_pure_builtin
and running the differential corpus showed all twenty-nine of these are
genuinely routed through it -- vm_preferred alone does not prove the
compiled path ran, since compilation declines for plenty of shapes -- so
with the divergence baseline empty these two implementations are already
known to agree over 3732 cases across four routes.
