+++ src/evaluator.rs
@@ -1220,7 +1220,7 @@ fn eval_compiled_inner(
                     arg, data, vars, ctx, shape, options, start_time,
                 )?);
             }
-            call_pure_builtin(name, &evaled_args, data, options)
+            crate::builtins::dispatch_pure(name, &evaled_args, data, options)
         }
 
         // Block: evaluate each expression in sequence, return the last value.
@@ -1697,17 +1697,6 @@ pub(crate) fn compiled_concat(lhs: JValue, rhs: JValue) -> Result<JValue, Evalua
     Ok(JValue::string(l + &r))
 }
 
-/// Entry point for the bytecode VM to call pure builtins by name.
-#[inline]
-pub(crate) fn call_pure_builtin_by_name(
-    name: &str,
-    args: &[JValue],
-    data: &JValue,
-    options: &EvaluatorOptions,
-) -> Result<JValue, EvaluatorError> {
-    call_pure_builtin(name, args, data, options)
-}
-
 // ──────────────────────────────────────────────────────────────────────────────
 // Phase 2: path compilation, builtin dispatch, and supporting helpers
 // ──────────────────────────────────────────────────────────────────────────────
@@ -2080,7 +2069,7 @@ pub(crate) fn normalize_lazy(value: &JValue) -> Result<JValue, EvaluatorError> {
 /// stringifies it to the literal text "undefined" rather than treating it as
 /// absent. Observable whenever the subject contains that text:
 /// `$substringBefore("xundefinedy", missing.x)` is "x", not the whole string.
-const JS_UNDEFINED_AS_STRING: &str = "undefined";
+pub(crate) const JS_UNDEFINED_AS_STRING: &str = "undefined";
 
 /// `$substring` with an undefined start, which JavaScript's arithmetic rather
 /// than any JSONata rule decides.
@@ -2091,7 +2080,7 @@ const JS_UNDEFINED_AS_STRING: &str = "undefined";
 /// undefined start yields the whole string or none of it depending only on
 /// whether a length was given. Nothing here can be derived from "treat the
 /// missing start as 0"; that gets the second case wrong.
-fn substring_with_undefined_start(s: &str, has_length: bool) -> JValue {
+pub(crate) fn substring_with_undefined_start(s: &str, has_length: bool) -> JValue {
     if has_length {
         JValue::string("")
     } else {
@@ -2101,9 +2090,9 @@ fn substring_with_undefined_start(s: &str, has_length: bool) -> JValue {
 
 /// Validate and coerce a builtin's arguments against its jsonata-js signature.
 ///
-/// One copy shared by all three dispatch paths -- the compiled/VM
-/// `call_pure_builtin`, the tree-walker's `evaluate_function_call`, and
-/// `call_builtin_with_values` for builtins passed by reference. They used to
+/// One copy shared by all three dispatch paths -- the compiled/VM path via
+/// `crate::builtins::dispatch_pure`, the tree-walker's `evaluate_function_call`,
+/// and `call_builtin_with_values` for builtins passed by reference. They used to
 /// hand-roll their own argument checks, which is why they disagreed with each
 /// other about `$round` and the `$substring*` family.
 ///
@@ -2165,548 +2154,6 @@ pub(crate) fn validate_builtin_args(
     }
 }
 
-/// Dispatch a pure builtin function call.
-///
-/// Replicates the tree-walker's evaluation for the subset of builtins in
-/// `COMPILABLE_BUILTINS`: no side effects, no lambdas, no context mutations.
-/// `data` is the current context value for implicit-argument insertion.
-fn call_pure_builtin(
-    name: &str,
-    args: &[JValue],
-    data: &JValue,
-    options: &EvaluatorOptions,
-) -> Result<JValue, EvaluatorError> {
-    use crate::functions;
-
-    // Apply implicit context insertion matching the tree-walker
-    let args_storage: Vec<JValue>;
-    let effective_args: &[JValue] = if args.is_empty() {
-        match name {
-            "string" => {
-                // $string() with a null/undefined context returns undefined, not "null".
-                // This mirrors the tree-walker's special case at the function-call site.
-                if data.is_undefined() || data.is_null() {
-                    return Ok(JValue::Undefined);
-                }
-                args_storage = vec![data.clone()];
-                &args_storage
-            }
-            "number" | "boolean" | "uppercase" | "lowercase" => {
-                args_storage = vec![data.clone()];
-                &args_storage
-            }
-            _ => args,
-        }
-    } else if args.len() == 1 {
-        match name {
-            "substringBefore" | "substringAfter" | "contains" | "split" => {
-                if matches!(data, JValue::String(_)) {
-                    args_storage = std::iter::once(data.clone())
-                        .chain(args.iter().cloned())
-                        .collect();
-                    &args_storage
-                } else {
-                    args
-                }
-            }
-            _ => args,
-        }
-    } else {
-        args
-    };
-
-    // Materialize top-level lazy args so every builtin sees plain Objects.
-    #[cfg(feature = "python")]
-    let lazy_storage: Vec<JValue>;
-    #[cfg(feature = "python")]
-    let effective_args: &[JValue] = if effective_args
-        .iter()
-        .any(|a| matches!(a, JValue::LazyPyDict(_)))
-    {
-        lazy_storage = effective_args
-            .iter()
-            .map(normalize_lazy)
-            .collect::<Result<Vec<_>, _>>()?;
-        &lazy_storage
-    } else {
-        effective_args
-    };
-
-    // Validate and coerce against the builtin's jsonata-js signature. This is
-    // where a lot of behaviour we used to hand-roll per function is actually
-    // specified: the `a` type "normally treats any value as a singleton array"
-    // (so `$reverse(1)` is `[1]` and `$count(null)` is 1), and `l` is distinct
-    // from `m`, so `$abs(null)` is a type error while a missing argument is not.
-    let sig_storage: Vec<JValue>;
-    let effective_args: &[JValue] = match validate_builtin_args(name, effective_args, data)? {
-        Some(coerced) => {
-            sig_storage = coerced;
-            &sig_storage
-        }
-        None => effective_args,
-    };
-
-    // Apply undefined propagation: if the first effective argument is Undefined
-    // and the function propagates undefined, return Undefined immediately.
-    //
-    // Deliberately *after* validation, which is the order jsonata-js works in:
-    // it has no such shortcut, it validates and then each function body checks
-    // its own first argument. The distinction is only visible when parameter 1
-    // is context-capable and a required parameter follows it -- `substring`
-    // (`<s-nn?:s>`) and `substringBefore`/`substringAfter` (`<s-s:s>`) are the
-    // whole set. There the lone undefined argument binds to parameter *2*, and
-    // parameter 1 comes from the context, so `$substring(missing.x)` against a
-    // non-string context is a T0411 rather than undefined. Running this guard
-    // first swallowed that, which is why the two engines disagreed: the
-    // tree-walker never had the shortcut.
-    if effective_args.first().is_some_and(JValue::is_undefined) && propagates_undefined(name) {
-        return Ok(JValue::Undefined);
-    }
-
-    match name {
-        // ── String functions ────────────────────────────────────────────
-        "string" => {
-            // Validate the optional prettify argument: must be a boolean.
-            let prettify = match effective_args.get(1) {
-                None => None,
-                Some(JValue::Bool(b)) => Some(*b),
-                Some(_) => {
-                    return Err(EvaluatorError::TypeError(
-                        "string() prettify parameter must be a boolean".to_string(),
-                    ))
-                }
-            };
-            let arg = effective_args.first().unwrap_or(&JValue::Null);
-            Ok(functions::string::string(arg, prettify)?)
-        }
-        "length" => match effective_args.first() {
-            Some(JValue::String(s)) => Ok(functions::string::length(s)?),
-            // Undefined input propagates (caught above by the undefined-propagation guard).
-            Some(JValue::Undefined) => Ok(JValue::Undefined),
-            // No argument: mirrors tree-walker "requires exactly 1 argument" (no error code,
-            // so the test framework accepts it against any expected T-code).
-            None => Err(EvaluatorError::EvaluationError(
-                "length() requires exactly 1 argument".to_string(),
-            )),
-            // null and any other non-string type → T0410
-            _ => Err(EvaluatorError::TypeError(
-                "T0410: Argument 1 of function length does not match function signature"
-                    .to_string(),
-            )),
-        },
-        "uppercase" => match effective_args.first() {
-            Some(JValue::String(s)) => Ok(functions::string::uppercase(s)?),
-            Some(JValue::Undefined) | None => Ok(JValue::Undefined),
-            _ => Err(EvaluatorError::TypeError(
-                "T0410: Argument 1 of function uppercase does not match function signature"
-                    .to_string(),
-            )),
-        },
-        "lowercase" => match effective_args.first() {
-            Some(JValue::String(s)) => Ok(functions::string::lowercase(s)?),
-            Some(JValue::Undefined) | None => Ok(JValue::Undefined),
-            _ => Err(EvaluatorError::TypeError(
-                "T0410: Argument 1 of function lowercase does not match function signature"
-                    .to_string(),
-            )),
-        },
-        "trim" => match effective_args.first() {
-            None | Some(JValue::Null | JValue::Undefined) => Ok(JValue::Null),
-            Some(JValue::String(s)) => Ok(functions::string::trim(s)?),
-            _ => Err(EvaluatorError::TypeError(
-                "trim() requires a string argument".to_string(),
-            )),
-        },
-        "substring" => {
-            if effective_args.len() < 2 {
-                return Err(EvaluatorError::EvaluationError(
-                    "substring() requires at least 2 arguments".to_string(),
-                ));
-            }
-            match (&effective_args[0], &effective_args[1]) {
-                (JValue::String(s), JValue::Number(start)) => {
-                    // Optional 3rd arg (length) must be a number if provided.
-                    let length = match effective_args.get(2) {
-                        None => None,
-                        Some(JValue::Number(l)) => Some(*l as i64),
-                        Some(_) => {
-                            return Err(EvaluatorError::TypeError(
-                                "T0410: Argument 3 of function substring does not match function signature"
-                                    .to_string(),
-                            ))
-                        }
-                    };
-                    Ok(functions::string::substring(s, *start as i64, length)?)
-                }
-                (JValue::String(s), JValue::Undefined) => {
-                    Ok(substring_with_undefined_start(s, effective_args.len() > 2))
-                }
-                _ => Err(EvaluatorError::TypeError(
-                    "T0410: Argument 1 of function substring does not match function signature"
-                        .to_string(),
-                )),
-            }
-        }
-        "substringBefore" => {
-            if effective_args.len() != 2 {
-                return Err(EvaluatorError::TypeError(
-                    "T0411: Context value is not a compatible type with argument 2 of function substringBefore".to_string(),
-                ));
-            }
-            match (&effective_args[0], &effective_args[1]) {
-                (JValue::String(s), JValue::String(sep)) => {
-                    Ok(functions::string::substring_before(s, sep)?)
-                }
-                (JValue::String(s), JValue::Undefined) => {
-                    Ok(functions::string::substring_before(s, JS_UNDEFINED_AS_STRING)?)
-                }
-                // Undefined propagates; null is a type error.
-                (JValue::Undefined, _) => Ok(JValue::Undefined),
-                _ => Err(EvaluatorError::TypeError(
-                    "T0410: Argument 1 of function substringBefore does not match function signature".to_string(),
-                )),
-            }
-        }
-        "substringAfter" => {
-            if effective_args.len() != 2 {
-                return Err(EvaluatorError::TypeError(
-                    "T0411: Context value is not a compatible type with argument 2 of function substringAfter".to_string(),
-                ));
-            }
-            match (&effective_args[0], &effective_args[1]) {
-                (JValue::String(s), JValue::String(sep)) => {
-                    Ok(functions::string::substring_after(s, sep)?)
-                }
-                (JValue::String(s), JValue::Undefined) => {
-                    Ok(functions::string::substring_after(s, JS_UNDEFINED_AS_STRING)?)
-                }
-                // Undefined propagates; null is a type error.
-                (JValue::Undefined, _) => Ok(JValue::Undefined),
-                _ => Err(EvaluatorError::TypeError(
-                    "T0410: Argument 1 of function substringAfter does not match function signature".to_string(),
-                )),
-            }
-        }
-        "contains" => {
-            if effective_args.len() != 2 {
-                return Err(EvaluatorError::EvaluationError(
-                    "contains() requires exactly 2 arguments".to_string(),
-                ));
-            }
-            // jsonata-js #809: $contains returns undefined when either argument
-            // (the string OR the pattern) is undefined.
-            if effective_args[0].is_undefined() || effective_args[1].is_undefined() {
-                return Ok(JValue::Undefined);
-            }
-            match &effective_args[0] {
-                JValue::Null => Ok(JValue::Null),
-                JValue::String(s) => Ok(functions::string::contains(s, &effective_args[1])?),
-                _ => Err(EvaluatorError::TypeError(
-                    "contains() requires a string as the first argument".to_string(),
-                )),
-            }
-        }
-        "split" => {
-            if effective_args.len() < 2 {
-                return Err(EvaluatorError::EvaluationError(
-                    "split() requires at least 2 arguments".to_string(),
-                ));
-            }
-            match &effective_args[0] {
-                JValue::Null | JValue::Undefined => Ok(JValue::Null),
-                JValue::String(s) => {
-                    // Validate the optional limit argument — must be a positive number.
-                    let limit = match effective_args.get(2) {
-                        None => None,
-                        Some(JValue::Number(n)) => {
-                            if *n < 0.0 {
-                                return Err(EvaluatorError::EvaluationError(
-                                    "D3020: Third argument of split function must be a positive number"
-                                        .to_string(),
-                                ));
-                            }
-                            Some(n.floor() as usize)
-                        }
-                        Some(_) => {
-                            return Err(EvaluatorError::TypeError(
-                                "split() limit must be a number".to_string(),
-                            ))
-                        }
-                    };
-                    Ok(functions::string::split(s, &effective_args[1], limit)?)
-                }
-                _ => Err(EvaluatorError::TypeError(
-                    "split() requires a string as the first argument".to_string(),
-                )),
-            }
-        }
-        "join" => {
-            if effective_args.is_empty() {
-                return Err(EvaluatorError::TypeError(
-                    "T0410: Argument 1 of function $join does not match function signature"
-                        .to_string(),
-                ));
-            }
-            match &effective_args[0] {
-                JValue::Null | JValue::Undefined => Ok(JValue::Null),
-                // Signature: <a<s>s?:s> — first arg must be an array of strings.
-                JValue::Bool(_) | JValue::Number(_) | JValue::Object(_) => {
-                    Err(EvaluatorError::TypeError(
-                        "T0412: Argument 1 of function $join must be an array of String"
-                            .to_string(),
-                    ))
-                }
-                #[cfg(feature = "python")]
-                JValue::LazyPyDict(_) => Err(EvaluatorError::TypeError(
-                    "T0412: Argument 1 of function $join must be an array of String".to_string(),
-                )),
-                JValue::Array(arr) => {
-                    // All elements must be strings.
-                    for item in arr.iter() {
-                        if !matches!(item, JValue::String(_)) {
-                            return Err(EvaluatorError::TypeError(
-                                "T0412: Argument 1 of function $join must be an array of String"
-                                    .to_string(),
-                            ));
-                        }
-                    }
-                    // Validate separator: must be a string if provided.
-                    let separator = match effective_args.get(1) {
-                        None | Some(JValue::Undefined) => None,
-                        Some(JValue::String(s)) => Some(&**s),
-                        Some(_) => {
-                            return Err(EvaluatorError::TypeError(
-                                "T0410: Argument 2 of function $join does not match function signature (expected String)"
-                                    .to_string(),
-                            ))
-                        }
-                    };
-                    Ok(functions::string::join(arr, separator)?)
-                }
-                JValue::String(s) => Ok(JValue::String(s.clone())),
-                _ => Err(EvaluatorError::TypeError(
-                    "T0412: Argument 1 of function $join must be an array of String".to_string(),
-                )),
-            }
-        }
-
-        // ── Numeric functions ───────────────────────────────────────────
-        "number" => match effective_args.first() {
-            Some(v) => Ok(functions::numeric::number(v)?),
-            None => Err(EvaluatorError::EvaluationError(
-                "number() requires at least 1 argument".to_string(),
-            )),
-        },
-        "floor" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Number(n)) => Ok(functions::numeric::floor(*n)?),
-            _ => Err(EvaluatorError::TypeError(
-                "floor() requires a number argument".to_string(),
-            )),
-        },
-        "ceil" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Number(n)) => Ok(functions::numeric::ceil(*n)?),
-            _ => Err(EvaluatorError::TypeError(
-                "ceil() requires a number argument".to_string(),
-            )),
-        },
-        "round" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Number(n)) => {
-                let precision = effective_args.get(1).and_then(|v| {
-                    if let JValue::Number(p) = v {
-                        Some(*p as i32)
-                    } else {
-                        None
-                    }
-                });
-                Ok(functions::numeric::round(*n, precision)?)
-            }
-            _ => Err(EvaluatorError::TypeError(
-                "round() requires a number argument".to_string(),
-            )),
-        },
-        "abs" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Number(n)) => Ok(functions::numeric::abs(*n)?),
-            _ => Err(EvaluatorError::TypeError(
-                "abs() requires a number argument".to_string(),
-            )),
-        },
-        "sqrt" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Number(n)) => Ok(functions::numeric::sqrt(*n)?),
-            _ => Err(EvaluatorError::TypeError(
-                "sqrt() requires a number argument".to_string(),
-            )),
-        },
-
-        // ── Aggregation functions ───────────────────────────────────────
-        "sum" => match effective_args.first() {
-            Some(v) if v.is_undefined() => Ok(JValue::Undefined),
-            None => Err(EvaluatorError::EvaluationError(
-                "sum() requires exactly 1 argument".to_string(),
-            )),
-            Some(JValue::Null) => Ok(JValue::Null),
-            Some(JValue::Array(arr)) => Ok(aggregation::sum(arr)?),
-            Some(JValue::Number(n)) => Ok(JValue::Number(*n)),
-            Some(other) => Ok(functions::numeric::sum(&[other.clone()])?),
-        },
-        "max" => match effective_args.first() {
-            Some(v) if v.is_undefined() => Ok(JValue::Undefined),
-            Some(JValue::Null) | None => Ok(JValue::Null),
-            Some(JValue::Array(arr)) => Ok(aggregation::max(arr)?),
-            Some(v @ JValue::Number(_)) => Ok(v.clone()),
-            _ => Err(EvaluatorError::TypeError(
-                "max() requires an array or number argument".to_string(),
-            )),
-        },
-        "min" => match effective_args.first() {
-            Some(v) if v.is_undefined() => Ok(JValue::Undefined),
-            Some(JValue::Null) | None => Ok(JValue::Null),
-            Some(JValue::Array(arr)) => Ok(aggregation::min(arr)?),
-            Some(v @ JValue::Number(_)) => Ok(v.clone()),
-            _ => Err(EvaluatorError::TypeError(
-                "min() requires an array or number argument".to_string(),
-            )),
-        },
-        "average" => match effective_args.first() {
-            Some(v) if v.is_undefined() => Ok(JValue::Undefined),
-            Some(JValue::Null) | None => Ok(JValue::Null),
-            Some(JValue::Array(arr)) => Ok(aggregation::average(arr)?),
-            Some(v @ JValue::Number(_)) => Ok(v.clone()),
-            _ => Err(EvaluatorError::TypeError(
-                "average() requires an array or number argument".to_string(),
-            )),
-        },
-        "count" => match effective_args.first() {
-            Some(v) if v.is_undefined() => Ok(JValue::from(0i64)),
-            Some(JValue::Null) | None => Ok(JValue::from(0i64)),
-            Some(JValue::Array(arr)) => Ok(functions::array::count(arr)?),
-            _ => Ok(JValue::from(1i64)),
-        },
-
-        // ── Boolean / logic ─────────────────────────────────────────────
-        "boolean" => match effective_args.first() {
-            Some(v) => Ok(functions::boolean::boolean(v)?),
-            None => Err(EvaluatorError::EvaluationError(
-                "boolean() requires 1 argument".to_string(),
-            )),
-        },
-        "not" => match effective_args.first() {
-            Some(v) => Ok(JValue::Bool(!compiled_is_truthy(v))),
-            None => Err(EvaluatorError::EvaluationError(
-                "not() requires 1 argument".to_string(),
-            )),
-        },
-
-        // ── Array functions ─────────────────────────────────────────────
-        "append" => {
-            if effective_args.len() != 2 {
-                return Err(EvaluatorError::EvaluationError(
-                    "append() requires exactly 2 arguments".to_string(),
-                ));
-            }
-            let first = &effective_args[0];
-            let second = &effective_args[1];
-            // Only a *missing* operand is skipped; an explicit null is a value
-            // and gets appended. Mirrors the tree-walker arm.
-            if matches!(second, JValue::Undefined) {
-                return Ok(first.clone());
-            }
-            if matches!(first, JValue::Undefined) {
-                return Ok(second.clone());
-            }
-            let arr = match first {
-                JValue::Array(a) => a.to_vec(),
-                other => vec![other.clone()],
-            };
-            let second_len = match second {
-                JValue::Array(a) => a.len(),
-                _ => 1,
-            };
-            check_sequence_length(arr.len() + second_len, options)?;
-            Ok(functions::array::append(&arr, second)?)
-        }
-        "reverse" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Array(arr)) => Ok(functions::array::reverse(arr)?),
-            _ => Err(EvaluatorError::TypeError(
-                "reverse() requires an array argument".to_string(),
-            )),
-        },
-        "distinct" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Array(arr)) if arr.len() > 1 => Ok(functions::array::distinct(arr)?),
-            // Non-array input, and arrays of length <= 1, pass through unchanged
-            // (jsonata-js functions.js: `if(!Array.isArray(arr) || arr.length <= 1) return arr;`)
-            Some(other) => Ok(other.clone()),
-        },
-
-        // ── Object functions ────────────────────────────────────────────
-        "keys" => match effective_args.first() {
-            Some(JValue::Null | JValue::Undefined) | None => Ok(JValue::Null),
-            Some(JValue::Lambda { .. } | JValue::Builtin { .. }) => Ok(JValue::Null),
-            Some(JValue::Object(obj)) => {
-                if obj.is_empty() {
-                    Ok(JValue::Null)
-                } else {
-                    let keys: Vec<JValue> = obj.keys().map(|k| JValue::string(k.clone())).collect();
-                    check_sequence_length(keys.len(), options)?;
-                    if keys.len() == 1 {
-                        Ok(keys.into_iter().next().unwrap())
-                    } else {
-                        Ok(JValue::array(keys))
-                    }
-                }
-            }
-            Some(JValue::Array(arr)) => {
-                let mut all_keys: Vec<JValue> = Vec::new();
-                for item in arr.iter() {
-                    let normalized_item = normalize_lazy(item)?;
-                    if let JValue::Object(obj) = &normalized_item {
-                        for key in obj.keys() {
-                            let k = JValue::string(key.clone());
-                            if !all_keys.contains(&k) {
-                                all_keys.push(k);
-                            }
-                        }
-                    }
-                }
-                if all_keys.is_empty() {
-                    Ok(JValue::Null)
-                } else if all_keys.len() == 1 {
-                    Ok(all_keys.into_iter().next().unwrap())
-                } else {
-                    check_sequence_length(all_keys.len(), options)?;
-                    Ok(JValue::array(all_keys))
-                }
-            }
-            _ => Ok(JValue::Null),
-        },
-        "merge" => match effective_args.len() {
-            0 => Err(EvaluatorError::EvaluationError(
-                "merge() requires at least 1 argument".to_string(),
-            )),
-            1 => match &effective_args[0] {
-                JValue::Array(arr) => Ok(functions::object::merge(arr)?),
-                JValue::Null | JValue::Undefined => Ok(JValue::Null),
-                JValue::Object(_) => Ok(effective_args[0].clone()),
-                _ => Err(EvaluatorError::TypeError(
-                    "merge() requires objects or an array of objects".to_string(),
-                )),
-            },
-            _ => Ok(functions::object::merge(effective_args)?),
-        },
-
-        _ => unreachable!(
-            "call_pure_builtin called with non-compilable builtin: {}",
-            name
-        ),
-    }
-}
-
 // ──────────────────────────────────────────────────────────────────────────────
 // End of CompiledExpr framework
 // ──────────────────────────────────────────────────────────────────────────────
@@ -2739,7 +2186,7 @@ pub(crate) fn propagates_undefined(name: &str) -> bool {
 
 /// Iterator-based numeric aggregation helpers.
 /// These avoid cloning values by iterating over references and extracting f64 values directly.
-mod aggregation {
+pub(crate) mod aggregation {
     use super::*;
 
     /// Iterate over all numeric values in a potentially nested array, yielding f64 values.
